+++ .env.test
@@ -13,6 +13,6 @@ TRELLO_BOARD_ID=222222222222222222222222
 FLASK_SKIP_DOTENV=True
 
 # Azure cosmos connection details
-CON_STRING='mongodb://fakemongo.com'
-DB_NAME='test_CosmosDBName'
-COLLECTION_NAME='test_CollectionName'
+CON_STRING=mongodb://fakemongo.com
+DB_NAME=test_CosmosDBName
+COLLECTION_NAME=test_CollectionName
